requirements.txt
discord.py
python-dotenv
mysql-connector-python
transformers
torch
numpy
requests
